requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.2.0
django-extensions==2.2.5
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-bulk==0.2.1
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
